+++ HISTORY.md
@@ -6,6 +6,7 @@
 * **Notable Changes:**
     * same blocks with empty variadic inputs compare as equal regardless of their arity
     * made "When I receive any messagge" non-thread-safe by default (again) to enable tail recursive broadcasts
+    * improved handling of user-defined errors and errors inside custom blocks 
 * **Notable Fixes:**
     * fixed storing the stage name(s)
     * removed distinction between number and string keys in "analyze" 
@@ -20,6 +21,7 @@
 * api: fixed variable binding when broadcasting through the API, thanks, Zak!
 * objects: fixed programmatically hiding palette blocks using the "hide variable" block, thanks, Zak!
 * threads, api: made "When I receive any messagge" non-thread-safe by default (again) to enable tail recursive broadcasts
+* threads, extensions: improved handling of user-defined errors and errors inside custom blocks 
 
 ### 2022-01-02
 * store: fixed storing the stage name(s)